pyproject.toml
[project]
name = "bit-babbler-0-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "libusb1>=3.3.1",
    "pyusb>=1.3.1",
]
